requirements.txt
fastapi
uvicorn
requests
tqdm
redis
celery
pydantic
python-dotenv
sentence-transformers
pymupdf
layoutparser
pillow
numpy
qdrant-client
boto3
pymupdf
pytesseract
Pillow
pdf2image
pdfplumber
google-genai
tiktoken
sqlalchemy
alembic
psycopg2-binary
python-jose 
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
langfuse
asyncpg
slowapi